remove bootstrap logic

+++ pkg/storage/database/database.go
@@ -30,7 +30,7 @@ type Database interface {
 	CreateTeam(name string) (*models.Team, error)
 	AddUserToTeam(userId uint, teamId uint) error
 
-	GetUser(int64) *models.User
+	GetUser(uint) *models.User
 	GetTeamId(userId uint) (uint, error)
 	CreateUser(email string, source string, details string) (*models.User, error)
